+++ petbook_fe/pages/api/parser/getResource.ts
